+++ core/XH.js
@@ -42,6 +42,7 @@ import parser from 'ua-parser-js';
 import {AppContainerModel} from '../appcontainer/AppContainerModel';
 import '../styles/XH.scss';
 import {ExceptionHandler} from './ExceptionHandler';
+import {HoistModel} from './HoistModel';
 import {RouterModel} from './RouterModel';
 
 const MIN_HOIST_CORE_VERSION = '14.0';
@@ -183,7 +184,6 @@ class XHClass {
     get isPhone()               {return this.uaParser.getDevice().type === 'mobile'}
     get isTablet()              {return this.uaParser.getDevice().type === 'tablet'}
     get isDesktop()             {return this.uaParser.getDevice().type === undefined}
-
 
     //---------------------------
     // Models
@@ -652,6 +652,23 @@ class XHClass {
     //---------------------------
     // Miscellaneous
     //---------------------------
+    /**
+     * Return a collection of models currently 'active' in this application.
+     *
+     * This will include all models that have not had the destroy() method
+     * called on them.  Models will be returned in creation order.
+     *
+     * @param {ModelSelector} [selector] - optional selector for filtering models.
+     * @returns {HoistModel[]}
+     */
+    getActiveModels(selector = '*') {
+        const ret = [];
+        HoistModel._activeModels.forEach(m => {
+            if (m.matchesSelector(selector, true)) ret.push(m);
+        });
+        return ret;
+    }
+
     /**
      * Resets user preferences and any persistent local application state, then reloads the app.
      */